requirements.txt
selenium
pytest
python-dotenv
pillow
